pyproject.toml
[project]
name = "loglensai-sidecar"
version = "0.1.0"
description = "High-performance Python sidecar for LogLensAi (DuckDB + Drain3 + aiohttp)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core storage & clustering
    "duckdb>=1.2.0",
    "drain3>=0.9.11",
    # HTTP server (dev mode)
    "aiohttp>=3.11.0",
    "aiohttp-cors>=0.7.0",
    # SSH tailing
    "paramiko>=3.5.0",
    # Data validation
    "pydantic>=2.10.0",
    # MCP Server
    "mcp>=1.26.0",
    "google-adk>=2.0.0a2",
    "datamodel-code-generator>=0.56.1",
    "openai>=2.32.0",
    "google-generativeai>=0.8.6",
    "google-genai>=1.69.0",
    "python-dotenv>=1.2.2",
    "langgraph>=1.1.9",
    "pydantic-ai>=1.31.0",
    "aiosqlite>=0.22.1",
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["B008", "N999", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["sidecar"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["sidecar/tests"]
